requirements.txt
typer[all]
pandas
numpy
matplotlib
seaborn
httpx
python-dotenv
google-generativeai
openai
pyyaml
tabulate
